Move Serialize context to it's own directory
To help keep serialization data separate from REST data, move it to a
different directory.
